% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norta.R
\name{determine_x}
\alias{determine_x}
\title{Determine x for NORTA}
\usage{
determine_x(z, x_cdf, tol = 10^-10)
}
\arguments{
\item{z}{value to evaluate cdf}

\item{x_cdf}{arbitrary cdf}

\item{tol}{numerical tolerance}
}
\value{
lower
}
\description{
Determine x for NORTA
}
