% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_ridge_functions.R
\name{fit_regressions}
\alias{fit_regressions}
\title{Fit and test regression model}
\usage{
fit_regressions(type, xtrain, xtest, ytrain, ytest, nfolds)
}
\arguments{
\item{type}{linear or logistic}

\item{xtrain}{training set covariate matrix}

\item{xtest}{test set covariate matrix}

\item{ytrain}{training set response}

\item{ytest}{test set response}

\item{nfolds}{number of folds to use in cross-validation}
}
\value{
coefficients of fitted models and test set errors (RMSE or accuracy)
}
\description{
Fit and test regression model
}
