% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_ridge_functions.R
\name{gen_data}
\alias{gen_data}
\title{Generate Data
Creates covariate matrix and response for linear or logistic models.}
\usage{
gen_data(
  type,
  n,
  p,
  rho,
  mu,
  stdev,
  spars,
  betas,
  train_size,
  sigma = NULL,
  skew = NULL
)
}
\arguments{
\item{type}{A character (linear, logistic) that specifies the type of model}

\item{n}{integer for number of observations, i.e., rows}

\item{p}{integer for number of predictors, i.e., columns, (not including intercept)}

\item{rho}{correlation between covariates if using multivariate normal matrix}

\item{mu}{means of predictors if using multivariate normal matrix}

\item{stdev}{standard deviation of predictors if using multivariate normal matrix}

\item{spars}{sparsity of covariate matrix by percentage}

\item{betas}{vector of true coefficients}

\item{train_size}{proportion of data to be used in training}

\item{sigma}{standard deviation of residuals for linear models}

\item{skew}{distribution of covariates. If null, multivariate normal is used.}
}
\value{
Generated data
}
\description{
Generate Data
Creates covariate matrix and response for linear or logistic models.
}
