% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norta.R
\name{generate_matrix_of_observations}
\alias{generate_matrix_of_observations}
\title{NORTA Observations
Generates covariate matrix using the normal-to-anything algorithm}
\usage{
generate_matrix_of_observations(corr_mat, cdf, num_rows = 1)
}
\arguments{
\item{corr_mat}{symmetric, positive definite correlation matrix}

\item{cdf}{cumulative distribution function of a random variable}

\item{num_rows}{number of observations, i.e., rows}
}
\value{
(nxp) matrix of observations
}
\description{
NORTA Observations
Generates covariate matrix using the normal-to-anything algorithm
}
