% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_ridge_functions.R
\name{run_simulations}
\alias{run_simulations}
\title{Runs simulation study to assess models}
\usage{
run_simulations(
  type,
  ntrials = 20,
  n = 1000,
  p = 10,
  rho = 0.5,
  mu = runif(p, 1, 5),
  stdev = runif(p, 0.2, 20),
  spars = 0.5,
  betas = runif(p, 0.01, 0.2),
  train_size = 0.75,
  nfolds = 5,
  sigma = NULL,
  skew = NULL
)
}
\arguments{
\item{type}{linear or logistic model}

\item{ntrials}{number of trials in simulation}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{rho}{correlation of predictors if using multivariate normal}

\item{mu}{means of predictors if using multivariate normal}

\item{stdev}{standard deviation of predictors if using multivariate normal}

\item{spars}{sparsity of covariate matrix as a proportion}

\item{betas}{true values of the coefficients}

\item{train_size}{proportion of data to use in training}

\item{nfolds}{number of folds for cross-validation}

\item{sigma}{standard deviation of the residuals for linear models}

\item{skew}{distribution of covariate matrix}
}
\value{
matrix of fitted coefficients, mean test set accuracies
}
\description{
Runs simulation study to assess models
}
